requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.3
django-material==1.4.1
python-decouple==3.1
pytz==2018.7
static3==0.7.0

gunicorn==19.9.0
psycopg2==2.7.6.1